2023: Default to u64 for return value

+++ 2023/src/day_00.rs
@@ -2,7 +2,7 @@ use advent_of_code::read_lines;
 use log::{debug, log_enabled, Level};
 use std::path::Path;
 
-pub fn part_one(input_path: &Path) -> i32 {
+pub fn part_one(input_path: &Path) -> u64 {
     if let Ok(lines) = read_lines(input_path) {
         for line in lines {
             if let Ok(result) = line {
@@ -14,7 +14,7 @@ pub fn part_one(input_path: &Path) -> i32 {
     0
 }
 
-pub fn part_two(input_path: &Path) -> i32 {
+pub fn part_two(input_path: &Path) -> u64 {
     if let Ok(lines) = read_lines(input_path) {
         for line in lines {
             if let Ok(result) = line {
